Update get_relations_by_pattern.sql
The get_relations_by_pattern macro does not work if the source tables in the database dont have a valid name (for example in our case the name included double --).
I added an additional parameter "quoting" and set the default value to ==false.
If the parameter quoting==true in the get_relations_by_pattern then it returns a table name from information_schema.tables in double quotes and that fixes the problem. 
If the parameter in a macro call is ommited, everything works as until now.